+++ lib/service/DeviceCalendar.dart
@@ -6,19 +6,18 @@ import 'package:illinois/service/AppDateTime.dart';
 import 'package:illinois/service/Auth2.dart';
 import 'package:illinois/service/Config.dart';
 import 'package:illinois/service/ExploreService.dart';
-
 import 'package:illinois/service/NotificationService.dart';
 import 'package:illinois/service/Service.dart';
+import 'package:illinois/service/Sports.dart';
 import 'package:illinois/service/Storage.dart';
 import 'package:illinois/utils/Utils.dart';
 import 'package:illinois/model/Event.dart' as ExploreEvent;
 
 class DeviceCalendar with Service implements NotificationsListener{
-
-  static const String notifyPromptPopupMessage            = "edu.illinois.rokwire.device_calendar.messaging.message.popup";
-  static const String notifyCalendarSelectionPopupMessage = "edu.illinois.rokwire.device_calendar.messaging.calendar_selection.popup";
-  static const String notifyPlaceEventMessage             = "edu.illinois.rokwire.device_calendar.messaging.place.event";
-  static const String showConsoleMessage                  = "edu.illinois.rokwire.debug_console.messaging.message";
+  static const String notifyPromptPopup            = "edu.illinois.rokwire.device_calendar.messaging.message.popup";
+  static const String notifyCalendarSelectionPopup = "edu.illinois.rokwire.device_calendar.messaging.calendar_selection.popup";
+  static const String notifyPlaceEvent             = "edu.illinois.rokwire.device_calendar.messaging.place.event";
+  static const String notifyShowConsoleMessage     = "edu.illinois.rokwire.device_calendar.console.debug.message";
 
   Calendar _defaultCalendar;
   List<Calendar> _deviceCalendars;
@@ -38,8 +37,15 @@ class DeviceCalendar with Service implements NotificationsListener{
   void createService() {
     NotificationService().subscribe(this, [
       Auth2UserPrefs.notifyFavoriteChanged,
-      DeviceCalendar.notifyPlaceEventMessage
+      DeviceCalendar.notifyPlaceEvent
     ]);
+  }
+
+  @override
+  Future<void> initService() async {
+    _deviceCalendarPlugin = new DeviceCalendarPlugin();
+    dynamic storedTable = Storage().calendarEventsTable ?? Map();
+    _calendarEventIdTable = storedTable!=null ? Map<String, String>.from(storedTable): Map();
   }
 
   @override
@@ -47,18 +53,18 @@ class DeviceCalendar with Service implements NotificationsListener{
     NotificationService().unsubscribe(this);
   }
 
-  Future<bool> _addEvent(DeviceCalendarEvent event) async{
+  Future<bool> _addEvent(_DeviceCalendarEvent event) async{
     //User prefs
     if(!canAddToCalendar){
       _debugMessage("Disabled");
       return false;
     }
 
-    if(_deviceCalendarPlugin == null){
-      bool initResult = await _initDeviceCalendarPlugin();
-      if(!initResult ?? true){
-        _debugMessage("Unable to init plugin");
-      }
+    //init check
+    bool initResult = await _loadDefaultCalendarIfNeeded();
+    if(!initResult ?? true){
+      _debugMessage("Unable to init plugin");
+      return false;
     }
     
     if(canShowPrompt){
@@ -69,51 +75,49 @@ class DeviceCalendar with Service implements NotificationsListener{
     return _placeCalendarEvent(event);
   }
 
-  Future<bool> _placeCalendarEvent(DeviceCalendarEvent event,) async{
+  Future<bool> _placeCalendarEvent(_DeviceCalendarEvent event) async{
     if(event == null)
       return false;
 
-    //PLUGIN
-    if(_deviceCalendarPlugin == null){
-      bool initResult = await _initDeviceCalendarPlugin();
-      if(!initResult ?? true){
-        _debugMessage("Unable to init plugin");
-      }
+    //init check
+    bool initResult = await _loadDefaultCalendarIfNeeded();
+    if(!initResult ?? true){
+      _debugMessage("Unable to init plugin");
+      return false;
     }
-    
-    _debugMessage("Add to calendar- id:${calendar?.id}, name:${calendar?.name}, accountName:${calendar?.accountName}, accountType:${calendar?.accountType}, isReadOnly:${calendar?.isReadOnly}, isDefault:${calendar?.isDefault},");
-    //PERMISSIONS
-    bool hasPermissions = await _requestPermissions();
 
-    _debugMessage("Has permissions: $hasPermissions");
+    _debugMessage("Add to calendar- id:${calendar?.id}, name:${calendar?.name}, accountName:${calendar?.accountName}, accountType:${calendar?.accountType}, isReadOnly:${calendar?.isReadOnly}, isDefault:${calendar?.isDefault},");
     //PLACE
-    if(hasPermissions && calendar!=null) {
+    if(calendar!=null) {
       final createEventResult = await _deviceCalendarPlugin.createOrUpdateEvent(event.toCalendarEvent(calendar?.id));
       if(createEventResult?.data!=null){
         _storeEventId(event.internalEventId, createEventResult?.data);
       }
 
-      _debugMessage("result.data: ${createEventResult.data}, result.errorMessages: ${createEventResult.errorMessages}");
+      _debugMessage("result.data: ${createEventResult?.data}, result.errorMessages: ${createEventResult?.errorMessages}");
 
       if(!createEventResult.isSuccess) {
         AppToast.show(createEventResult?.data ?? createEventResult?.errorMessages ?? "Unable to save Event to calendar");
         print(createEventResult?.errorMessages);
         return false;
       }
+    } else {
+      _debugMessage("calendar is missing");
     }
-    
+
+    _debugMessage("added");
     return true;
   }
 
-  Future<bool> deleteEvent(DeviceCalendarEvent event) async{
+  Future<bool> _deleteEvent(_DeviceCalendarEvent event) async{
     if(event == null)
       return false;
 
-    if(_deviceCalendarPlugin == null){
-      bool initResult = await _initDeviceCalendarPlugin();
-      if(!initResult ?? true){
-        _debugMessage("Unable to init plugin");
-      }
+    //init check
+    bool initResult = await _loadDefaultCalendarIfNeeded();
+    if(!initResult ?? true){
+      _debugMessage("Unable to init plugin");
+      return false;
     }
 
     String eventId = event?.internalEventId != null && _calendarEventIdTable!= null ? _calendarEventIdTable[event?.internalEventId] : null;
@@ -129,19 +133,21 @@ class DeviceCalendar with Service implements NotificationsListener{
     }
     return deleteEventResult?.isSuccess;
   }
-
-  Future<bool> _initDeviceCalendarPlugin() async{
-    _deviceCalendarPlugin = new DeviceCalendarPlugin();
-    dynamic storedTable = Storage().calendarEventsTable ?? Map();
-    _calendarEventIdTable = storedTable!=null ? Map<String, String>.from(storedTable): Map();
+  
+  Future<bool> _loadDefaultCalendarIfNeeded() async{
+    if(calendar!=null)
+      return true;
+    
     return await _loadCalendars();
   }
 
   Future<bool> _loadCalendars() async {
     bool hasPermissions = await _requestPermissions();
     if(!hasPermissions) {
+      _debugMessage("No Calendar permissions");
       return false;
     }
+    _debugMessage("Has permissions");
     final calendarsResult = await _deviceCalendarPlugin.retrieveCalendars();
     List<Calendar> calendars = calendarsResult.data;
     _deviceCalendars = calendars!=null && calendars.isNotEmpty? calendars.where((Calendar calendar) => calendar.isReadOnly == false)?.toList() : null;
@@ -152,7 +158,7 @@ class DeviceCalendar with Service implements NotificationsListener{
         return true;
       }
     }
-
+    _debugMessage("No Calendars");
     return false;
   }
 
@@ -184,11 +190,11 @@ class DeviceCalendar with Service implements NotificationsListener{
   }
 
   void _debugMessage(String msg){
-    NotificationService().notify(DeviceCalendar.showConsoleMessage, msg);
+    NotificationService().notify(DeviceCalendar.notifyShowConsoleMessage, msg);
   }
 
   void _processEvents(dynamic event){
-    DeviceCalendarEvent deviceCalendarEvent = DeviceCalendarEvent.from(event);
+    _DeviceCalendarEvent deviceCalendarEvent = _DeviceCalendarEvent.from(event);
     if(deviceCalendarEvent==null)
       return;
 
@@ -196,22 +202,24 @@ class DeviceCalendar with Service implements NotificationsListener{
       _addEvent(deviceCalendarEvent);
     }
     else {
-      deleteEvent(deviceCalendarEvent);
+      _deleteEvent(deviceCalendarEvent);
     }
   }
 
-  void _promptPermissionDialog(DeviceCalendarEvent event) {
-    NotificationService().notify(DeviceCalendar.notifyCalendarSelectionPopupMessage, {"event": event});
+  void _promptPermissionDialog(_DeviceCalendarEvent event) {
+    NotificationService().notify(DeviceCalendar.notifyCalendarSelectionPopup, {"event": event});
   }
 
   @override
   void onNotification(String name, param) {
     if(name == Auth2UserPrefs.notifyFavoriteChanged){
       _processEvents(param);
-    } else if(name == DeviceCalendar.notifyPlaceEventMessage){
+    }
+    else if(name == DeviceCalendar.notifyPlaceEvent){
       if(param!=null && param is Map){
-        DeviceCalendarEvent event = param["event"];
+        _DeviceCalendarEvent event = param["event"];
         Calendar calendarSelection = param["calendar"];
+
         if(calendarSelection!=null){
           _selectedCalendar = calendarSelection;
         }
@@ -237,58 +245,59 @@ class DeviceCalendar with Service implements NotificationsListener{
   }
 }
 
-class DeviceCalendarEvent {
+class _DeviceCalendarEvent {
   String internalEventId;
   String title;
   String deepLinkUrl;
   DateTime startDate;
   DateTime endDate;
 
-  DeviceCalendarEvent({this.internalEventId, this.title, this.deepLinkUrl, this.startDate, this.endDate});
+  _DeviceCalendarEvent({this.internalEventId, this.title, this.deepLinkUrl, this.startDate, this.endDate});
 
-  factory DeviceCalendarEvent.from(dynamic data){
+  factory _DeviceCalendarEvent.from(dynamic data){
     if(data==null)
       return null;
 
     if(data is ExploreEvent.Event){
-      return DeviceCalendarEvent.fromEvent(data);
+      return _DeviceCalendarEvent.fromEvent(data);
     }
     else if (data is Game){
-      return DeviceCalendarEvent.fromGame(data);
+      return _DeviceCalendarEvent.fromGame(data);
     }
 
     return null;
   }
 
-  factory DeviceCalendarEvent.fromEvent(ExploreEvent.Event event){
+  factory _DeviceCalendarEvent.fromEvent(ExploreEvent.Event event){
     if(event==null)
       return null;
 
-    return DeviceCalendarEvent(title: event.title, internalEventId: event.id, startDate: event.startDateLocal, endDate: event.endDateLocal,
+    return _DeviceCalendarEvent(title: event.title, internalEventId: event.id, startDate: event.startDateLocal,
+        endDate: event.endDateLocal,
         deepLinkUrl: "${ExploreService.EVENT_URI}?event_id=${event.id}");
   }
 
-  factory DeviceCalendarEvent.fromGame(Game game){
+  factory _DeviceCalendarEvent.fromGame(Game game){
     if(game==null)
       return null;
 
-    return DeviceCalendarEvent(title: game.title, internalEventId: game.id, startDate: game.dateTimeUniLocal,
+    return _DeviceCalendarEvent(title: game.title, internalEventId: game.id, startDate: game.dateTimeUniLocal,
         endDate:  AppDateTime().getUniLocalTimeFromUtcTime(game.endDateTimeUtc),
-        deepLinkUrl: "${ExploreService.EVENT_URI}?game_id=${game.id}");
+        deepLinkUrl: "${Sports.GAME_URI}?game_id=${game.id}%26sport=${game.sport?.shortName}");
   }
 
   Event toCalendarEvent(String calendarId){
     Event calendarEvent = Event(calendarId);
     calendarEvent.title = title ?? "";
+
     if (startDate != null) {
       calendarEvent.start = startDate;
     }
+
     if (endDate != null) {
       calendarEvent.end = endDate;
     } else {
-      calendarEvent.end = DateTime(
-        startDate.year, startDate.month,
-        startDate.day, 23, 59,);
+      calendarEvent.end = AppDateTime().localEndOfDay(startDate);
     }
 
     calendarEvent.description = _constructRedirectLinkUrl(deepLinkUrl);
@@ -297,11 +306,9 @@ class DeviceCalendarEvent {
   }
 
   static String _constructRedirectLinkUrl(String url){
-
     Uri assetsUri = Uri.parse(Config().assetsUrl);
     String redirectUrl = assetsUri!= null ? "${assetsUri.scheme}://${assetsUri.host}/html/redirect.html" : null;
 
     return AppString.isStringNotEmpty(redirectUrl) ? "$redirectUrl?target=$url" : url;
   }
-
 }